feat(sankey): move computing out of the rendering component
